refactor: rename apply pkg
