+++ pkg/apply/gen_test.go
@@ -1,0 +1,49 @@
+/*
+Copyright 2023 cuisongliu@qq.com.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package apply
+
+import (
+	"reflect"
+	"testing"
+)
+
+func TestNewClusterFromGenArgs(t *testing.T) {
+	type args struct {
+		imageName []string
+		args      *RunArgs
+	}
+	tests := []struct {
+		name    string
+		args    args
+		want    []interface{}
+		wantErr bool
+	}{
+		// TODO: Add test cases.
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := NewClusterFromGenArgs(tt.args.imageName, tt.args.args)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("NewClusterFromGenArgs() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("NewClusterFromGenArgs() got = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
