+++ pkg/apply/run.go
@@ -16,14 +16,11 @@ package apply
 
 import (
 	"fmt"
-	"os"
-	"path/filepath"
 	"strconv"
 
 	"github.com/labring/sealos/pkg/apply/applydrivers"
 	"github.com/labring/sealos/pkg/clusterfile"
 	"github.com/labring/sealos/pkg/constants"
-	"github.com/labring/sealos/pkg/runtime"
 	"github.com/labring/sealos/pkg/ssh"
 	v2 "github.com/labring/sealos/pkg/types/v1beta1"
 	"github.com/labring/sealos/pkg/utils/iputils"
@@ -35,29 +32,6 @@ type ClusterArgs struct {
 	cluster     *v2.Cluster
 	hosts       []v2.Host
 	clusterName string
-}
-
-func NewClusterFromArgs(imageName []string, args *RunArgs) ([]interface{}, error) {
-	cluster := initCluster(args.ClusterName)
-	c := &ClusterArgs{
-		clusterName: args.ClusterName,
-		cluster:     cluster,
-	}
-	if err := c.SetClusterRunArgs(imageName, args); err != nil {
-		return nil, err
-	}
-	kubeadmcfg := &runtime.KubeadmConfig{}
-	if err := kubeadmcfg.Merge(""); err != nil {
-		return nil, err
-	}
-	// todo: only generate configurations of the corresponding components by passing parameters
-	return []interface{}{c.cluster,
-		kubeadmcfg.InitConfiguration,
-		kubeadmcfg.ClusterConfiguration,
-		kubeadmcfg.JoinConfiguration,
-		kubeadmcfg.KubeProxyConfiguration,
-		kubeadmcfg.KubeletConfiguration,
-	}, nil
 }
 
 func NewApplierFromArgs(imageName []string, args *RunArgs) (applydrivers.Interface, error) {
@@ -81,51 +55,13 @@ func NewApplierFromArgs(imageName []string, args *RunArgs) (applydrivers.Interfa
 		clusterName: cluster.Name,
 		cluster:     cluster,
 	}
-	if err = c.SetClusterRunArgs(imageName, args); err != nil {
+	if err = c.runArgs(imageName, args); err != nil {
 		return nil, err
 	}
 	return applydrivers.NewDefaultApplier(c.cluster, cf, imageName)
 }
 
-func NewApplierFromFile(path string, args *Args) (applydrivers.Interface, error) {
-	if !filepath.IsAbs(path) {
-		pa, err := os.Getwd()
-		if err != nil {
-			return nil, err
-		}
-		path = filepath.Join(pa, path)
-	}
-	Clusterfile := clusterfile.NewClusterFile(path,
-		clusterfile.WithCustomValues(args.Values),
-		clusterfile.WithCustomSets(args.Sets),
-		clusterfile.WithCustomEnvs(args.CustomEnv),
-		clusterfile.WithCustomConfigFiles(args.CustomConfigFiles),
-	)
-	if err := Clusterfile.Process(); err != nil {
-		return nil, err
-	}
-	cluster := Clusterfile.GetCluster()
-	if cluster.Name == "" {
-		return nil, fmt.Errorf("cluster name cannot be empty, make sure %s file is correct", path)
-	}
-
-	localpath := constants.Clusterfile(cluster.Name)
-	cf := clusterfile.NewClusterFile(localpath)
-	err := cf.Process()
-	if err != nil && err != clusterfile.ErrClusterFileNotExists {
-		return nil, err
-	}
-	currentCluster := cf.GetCluster()
-
-	return &applydrivers.Applier{
-		ClusterDesired: cluster,
-		ClusterFile:    Clusterfile,
-		ClusterCurrent: currentCluster,
-		RunNewImages:   nil,
-	}, nil
-}
-
-func (r *ClusterArgs) SetClusterRunArgs(imageList []string, args *RunArgs) error {
+func (r *ClusterArgs) runArgs(imageList []string, args *RunArgs) error {
 	if args.Cluster.ClusterName == "" {
 		return fmt.Errorf("cluster name can not be empty")
 	}
@@ -187,49 +123,6 @@ func (r *ClusterArgs) SetClusterRunArgs(imageList []string, args *RunArgs) error
 		r.setHostWithIpsPort(nodes, []string{v2.NODE, GetHostArch(sshClient, nodes[0])})
 	}
 	r.cluster.Spec.Hosts = append(r.cluster.Spec.Hosts, r.hosts...)
-	logger.Debug("cluster info: %v", r.cluster)
-	return nil
-}
-
-func (r *ClusterArgs) SetClusterResetArgs(args *ResetArgs) error {
-	if args.Cluster.ClusterName == "" {
-		return fmt.Errorf("cluster name can not be empty")
-	}
-	if err := PreProcessIPList(args.Cluster); err != nil {
-		return err
-	}
-	if args.fs != nil {
-		if args.fs.Changed("user") || r.cluster.Spec.SSH.User == "" {
-			r.cluster.Spec.SSH.User = args.SSH.User
-		}
-		if args.fs.Changed("pk") || r.cluster.Spec.SSH.Pk == "" {
-			r.cluster.Spec.SSH.Pk = args.SSH.Pk
-		}
-		if args.fs.Changed("pk-passwd") || r.cluster.Spec.SSH.PkPasswd == "" {
-			r.cluster.Spec.SSH.PkPasswd = args.SSH.PkPassword
-		}
-		if args.fs.Changed("port") || r.cluster.Spec.SSH.Port == 0 {
-			r.cluster.Spec.SSH.Port = args.SSH.Port
-		}
-		if args.fs.Changed("passwd") || r.cluster.Spec.SSH.Passwd == "" {
-			r.cluster.Spec.SSH.Passwd = args.SSH.Password
-		}
-	}
-
-	if len(args.Cluster.Masters) > 0 {
-		masters := stringsutil.SplitRemoveEmpty(args.Cluster.Masters, ",")
-		nodes := stringsutil.SplitRemoveEmpty(args.Cluster.Nodes, ",")
-		r.hosts = []v2.Host{}
-
-		clusterSSH := r.cluster.GetSSH()
-		sshClient := ssh.NewSSHClient(&clusterSSH, true)
-
-		r.setHostWithIpsPort(masters, []string{v2.MASTER, GetHostArch(sshClient, masters[0])})
-		if len(nodes) > 0 {
-			r.setHostWithIpsPort(nodes, []string{v2.NODE, GetHostArch(sshClient, nodes[0])})
-		}
-		r.cluster.Spec.Hosts = r.hosts
-	}
 	logger.Debug("cluster info: %v", r.cluster)
 	return nil
 }
