+++ pkg/apply/run_test.go
@@ -25,60 +25,6 @@ import (
 	v2 "github.com/labring/sealos/pkg/types/v1beta1"
 )
 
-func Test_NewApplierFromFile(t *testing.T) {
-	type args struct {
-		customEnv []string
-		sets      []string
-		values    []string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "success",
-			args: args{
-				customEnv: []string{
-					"SSH_PASSWORD=s3cret",
-				},
-				sets: []string{
-					"clusterName=default",
-				},
-				values: []string{"../clusterfile/testdata/example.values.yaml"},
-			},
-			wantErr: false,
-		},
-		{
-			name: "fail",
-			args: args{
-				customEnv: []string{
-					"SSH_PASSWORD=s3cret",
-				},
-				sets: []string{
-					"clusterName=''",
-				},
-				values: []string{"../clusterfile/testdata/example.values.yaml"},
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, err := NewApplierFromFile("../clusterfile/testdata/clusterfile.yaml",
-				&Args{
-					Values:    tt.args.values,
-					Sets:      tt.args.sets,
-					CustomEnv: tt.args.customEnv,
-				})
-			t.Log(err)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("NewApplierFromFile(string, ...OptionFunc) error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
 func TestClusterArgs_SetClusterRunArgs(t *testing.T) {
 	type fields struct {
 		cluster     *v2.Cluster
@@ -141,8 +87,8 @@ func TestClusterArgs_SetClusterRunArgs(t *testing.T) {
 				hosts:       tt.fields.hosts,
 				clusterName: tt.fields.clusterName,
 			}
-			if err := r.SetClusterRunArgs(tt.args.imageList, tt.args.runArgs); (err != nil) != tt.wantErr {
-				t.Errorf("SetClusterRunArgs() error = %v, wantErr %v", err, tt.wantErr)
+			if err := r.runArgs(tt.args.imageList, tt.args.runArgs); (err != nil) != tt.wantErr {
+				t.Errorf("runArgs() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
 	}
@@ -199,60 +145,6 @@ func TestNewApplierFromArgs(t *testing.T) {
 			}
 			if !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("NewApplierFromArgs() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestNewApplierFromFile(t *testing.T) {
-	type args struct {
-		path string
-		args *Args
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    applydrivers.Interface
-		wantErr bool
-	}{
-		// TODO: Add test cases.
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := NewApplierFromFile(tt.args.path, tt.args.args)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("NewApplierFromFile() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("NewApplierFromFile() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestNewClusterFromArgs(t *testing.T) {
-	type args struct {
-		imageName []string
-		args      *RunArgs
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *v2.Cluster
-		wantErr bool
-	}{
-		// TODO: Add test cases.
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := NewClusterFromArgs(tt.args.imageName, tt.args.args)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("NewClusterFromArgs() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("NewClusterFromArgs() got = %v, want %v", got, tt.want)
 			}
 		})
 	}
